messaging/serializers.py
from rest_framework import serializers

from .models import Notification


class NotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = Notification
        fields = ["id", "ref_id", "ref_model", "message", "read_status", "timestamp"]
        read_only_fields = ["timestamp", "message"]
