requirements.txt
django~=4.2.6
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.3.1
daphne~=4.0.0
channels~=4.0.0
redis~=5.0.1
hiredis~=2.3.2
celery~=5.3.5
#mysql-connector-python
#psycopg2-binary
pyjwt~=2.8.0

requests~=2.31.0
urllib3==1.26.6
gTTS~=2.4.0
googletrans~=4.0.0rc1
googlemaps~=4.10.0
gunicorn
google-search-results
pipdeptree
